requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
sqlmodel==0.0.22
pgvector==0.3.5
pydantic==2.9.2
psycopg2-binary==2.9.9
alembic==1.12.0
celery==5.3.6
redis==5.0.1
sqladmin==0.19.0
python-dotenv
aiofiles==23.2.1
aiohttp==3.10.10
anthropic==0.36.1
openai==1.51.2
numpy==1.26.4
python-socketio==5.11.4
python-engineio==4.10.1
pyjwt==2.9.0